+++ plugin/lessonmark/export_pdf.php
@@ -1,0 +1,32 @@
+<?php
+// This file is part of Moodle - https://moodle.org/
+//
+// Moodle is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+
+/**
+ * Downloads a PDF of the saved LessonMark activity.
+ *
+ * @package   mod_lessonmark
+ * @copyright 2026 Hiroshi Ozeki
+ * @license   https://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
+ */
+
+require_once(__DIR__ . '/../../config.php');
+
+$id = required_param('id', PARAM_INT);
+$cm = get_coursemodule_from_id('lessonmark', $id, 0, false, MUST_EXIST);
+$lessonmark = $DB->get_record('lessonmark', ['id' => $cm->instance], '*', MUST_EXIST);
+$course = $DB->get_record('course', ['id' => $cm->course], '*', MUST_EXIST);
+
+require_course_login($course, true, $cm);
+$context = context_module::instance($cm->id);
+require_capability('mod/lessonmark:edit', $context);
+require_sesskey();
+
+$name = format_string($lessonmark->name, true, ['context' => $context]);
+$filename = \mod_lessonmark\local\pdf_exporter::export_filename($name);
+$content = (new \mod_lessonmark\local\pdf_exporter())->generate($lessonmark, $course, $context);
+send_temp_file($content, $filename, true);